Refined user data models to fit better to actual data

+++ app/src/main/java/me/stammberger/starcitizencompact/models/user/Performance.java
@@ -10,7 +10,7 @@ public class Performance {
 
     @SerializedName("network_io_time")
     @Expose
-    public Double networkIoTime;
+    public Integer networkIoTime;
     @SerializedName("processing_time")
     @Expose
     public Double processingTime;
